# perf(gatsby): cache babel config items