--- a/packages/gatsby/src/utils/babel-loader-helpers.js
+++ b/packages/gatsby/src/utils/babel-loader-helpers.js
@@ -21,10 +21,23 @@ const getCustomOptions = stage => {
   return pluginBabelConfig.stages[stage].options
 }
 
-const prepareOptions = (babel, options = {}, resolve = require.resolve) => {
-  const pluginBabelConfig = loadCachedConfig()
+/**
+ * https://babeljs.io/docs/en/babel-core#createconfigitem
+ * If this function is called multiple times for a given plugin,
+ * Babel will call the plugin's function itself multiple times.
+ * If you have a clear set of expected plugins and presets to inject,
+ * pre-constructing the config items would be recommended.
+ */
+const configItemsMemoCache = new Map()
 
+const prepareOptions = (babel, options = {}, resolve = require.resolve) => {
   const { stage, reactRuntime } = options
+
+  if (configItemsMemoCache.has(stage)) {
+    return configItemsMemoCache.get(stage)
+  }
+
+  const pluginBabelConfig = loadCachedConfig()
 
   // Required plugins/presets
   const requiredPlugins = [
@@ -95,13 +108,17 @@ const prepareOptions = (babel, options = {}, resolve = require.resolve) => {
     )
   })
 
-  return [
+  const toReturn = [
     reduxPresets,
     reduxPlugins,
     requiredPresets,
     requiredPlugins,
     fallbackPresets,
   ]
+
+  configItemsMemoCache.set(stage, toReturn)
+
+  return toReturn
 }
 
 const addRequiredPresetOptions = (