@KyleAMathews 2020-12-23
This PR is similar https://github.com/gatsbyjs/gatsby/pull/27974